+++ docs/tr1/CHANGELOG.md
@@ -17,6 +17,7 @@
 - fixed the bear AI fix option being applied in the Vilcabamba demo (#2559, regression from 4.8)
 - fixed extremely large item quantities crashing the game (#2497, regression from 0.3)
 - fixed Lara's meshes not resetting after using the fly cheat (#2565, #2572, regressions from 4.8)
+- fixed the select level feature not giving Lara her items (#2617, regression from 4.8)
 - fixed guns appearing in Lara's hands if the draw input is held when unarmed and while picking up a gun item (#2577, regressions from 0.8/4.3)
 - fixed being able to play with Lara invisible after using the explosion cheat then the fly cheat (#2584, regression from 4.8)
 - fixed the `/pos` command not working in cutscenes